generate-args: fix signed overflow in FT_SRANGE span calc
The FT_SRANGE fill arm computed the span as

    span = (uint64_t) ((int64_t) hi - (int64_t) lo + 1);

The subtraction and +1 are evaluated in signed int64_t before the
outer cast.  For a field that legitimately covers the full signed
range (e.g. lo == LONG_MIN, hi == LONG_MAX) the true span exceeds
LONG_MAX and the signed arithmetic invokes undefined behavior.

The preceding hi <= lo guard already excludes the empty and inverted
cases, so hi > lo holds at this point.  Switch the calculation to
unsigned:

    span = (uint64_t) hi - (uint64_t) lo + 1;

Two's-complement unsigned subtraction is fully defined and yields the
correct count of values in [lo, hi] for any hi >= lo, including the
full signed range.

Currently latent because no field definition emits FT_SRANGE with a
maximal range yet, but harden it now so the first such field does not
trip UBSan or get miscompiled at higher optimisation levels.

+++ generate-args.c
@@ -2075,7 +2075,7 @@ static void struct_fill_passes(unsigned char *buf, unsigned int size,
 				fill_field_raw(buf, f);
 				break;
 			}
-			span = (uint64_t) ((int64_t) hi - (int64_t) lo + 1);
+			span = (uint64_t) hi - (uint64_t) lo + 1;
 			v = (int64_t) lo +
 			    (int64_t) rnd_modulo_u64(span);
 			write_field_uint(buf, f, (uint64_t) v);
